Move GCHandle from System.Runtime to System.Runtime.InteropServices.PInvoke
